Add files via upload
This project aims to detect spam comments on YouTube using the Naive Bayes algorithm, specifically the BernoulliNB model. Spam detection is important for maintaining the quality and integrity of online platforms, and YouTube is no exception. In this project, we use a dataset containing YouTube comments and a binary label (Spam or Not Spam) to build a machine learning model for classification.